review: address Copilot feedback on PR #84 (5 comments)
1. computeFilteredDiff now passes --no-color and --no-ext-diff to git
   so a user's local color.ui=always or diff.external=delta config
   doesn't inject ANSI escapes / replace the canonical unified diff.

2. Updated computeFilteredDiff's failure-mode comment: removed the
   stale "empty diff" case (the function always falls back to the
   full diff when fileGlobs is empty; SKILL.md mentions this fallback
   explicitly).

3. Replaced the inline parseLeafFrontmatter / extractFileGlobs helpers
   with a shared lib at scripts/lib/leaf-frontmatter.mjs. Docstrings
   in the shared lib accurately describe the returned shapes (raw YAML
   text, not parsed object). run-review.mjs now imports
   splitFrontmatter + extractFileGlobs from the shared lib.

4. SKILL.md's dispatch_specialists section now explicitly mentions the
   full-diff fallback for the rare leaf without file_globs.

Note: verify-coverage.mjs still has its own readLeafGlobs with caching
and symlink hardening — its tri-state (null vs []) semantics are
load-bearing for broad-credit decisions there. Consolidation is
tracked separately; this PR keeps the per-PR scope tight.
